requirements.txt
python==3.8.20
pandas==1.5.3
numpy==1.22.4
scikit-learn==1.0.2
lightgbm==3.3.5
lightfm==1.17
flask==2.3.2
jupyter==1.0.0
matplotlib==3.7.1
seaborn==0.12.2
scipy==1.7.3
plotly==5.13.1
PyYAML==5.4.1
tqdm==4.65.0
requests==2.31.0
Cython==0.29.32